+++ io.h
@@ -2,7 +2,7 @@
 #define IO_H
 #include <string>
 
-void printMatrix(double M[3][50], int rows, int cols, std::string title);
-void msgOut(double M[3][50], int rows, int cols);
+void printMatrix(int M[3][50], int rows, int cols, std::string title);
+void msgOut(int M[3][50], int rows, int cols);
 
 #endif
